Moved types for StepButton in UI
