+++ src/features/identityCheck/pages/helpers/computeIdentificationMethod.native.test.ts
@@ -1,6 +1,6 @@
 import { IdentityCheckMethod } from 'api/gen'
 import { computeIdentificationMethod } from 'features/identityCheck/pages/helpers/computeIdentificationMethod'
-import { SubscriptionScreen } from 'features/identityCheck/types'
+import { SubscriptionScreen } from 'ui/components/StepButton/types'
 
 const educonnectFlowFirstScreen: SubscriptionScreen = 'EduConnectForm'
 const ubbleFlowFirstScreen: SubscriptionScreen = 'SelectIDOrigin'
