+++ src/features/identityCheck/pages/helpers/mapStepsDetails.native.test.tsx
@@ -9,12 +9,8 @@ import {
 import { IconRetryStep } from 'features/identityCheck/components/IconRetryStep'
 import { IconStepDone } from 'features/identityCheck/components/IconStepDone'
 import { mapStepsDetails } from 'features/identityCheck/pages/helpers/mapStepsDetails'
-import {
-  IdentityCheckStep,
-  StepButtonState,
-  StepConfig,
-  StepDetails,
-} from 'features/identityCheck/types'
+import { IdentityCheckStep } from 'features/identityCheck/types'
+import { StepButtonState, StepConfig, StepDetails } from 'ui/components/StepButton/types'
 import { BicolorIdCard } from 'ui/svg/icons/BicolorIdCard'
 import { BicolorLegal } from 'ui/svg/icons/BicolorLegal'
 import { BicolorProfile } from 'ui/svg/icons/BicolorProfile'
