+++ src/features/identityCheck/pages/helpers/useStepperInfo.tsx
@@ -6,8 +6,9 @@ import { IconRetryStep } from 'features/identityCheck/components/IconRetryStep'
 import { IconStepDone } from 'features/identityCheck/components/IconStepDone'
 import { computeIdentificationMethod } from 'features/identityCheck/pages/helpers/computeIdentificationMethod'
 import { mapStepsDetails } from 'features/identityCheck/pages/helpers/mapStepsDetails'
-import { IdentityCheckStep, StepConfig, StepDetails } from 'features/identityCheck/types'
+import { IdentityCheckStep } from 'features/identityCheck/types'
 import { theme } from 'theme'
+import { StepConfig, StepDetails } from 'ui/components/StepButton/types'
 import { BicolorIdCard } from 'ui/svg/icons/BicolorIdCard'
 import { BicolorLegal } from 'ui/svg/icons/BicolorLegal'
 import { BicolorProfile } from 'ui/svg/icons/BicolorProfile'
