+++ src/features/identityCheck/types.ts
@@ -1,6 +1,3 @@
-import { SubscriptionRootStackParamList } from 'features/navigation/RootNavigator/types'
-import { IconInterface } from 'ui/svg/icons/types'
-
 export enum DeprecatedIdentityCheckStep {
   PHONE_VALIDATION = 'phone_validation',
   PROFILE = 'profile',
@@ -15,27 +12,6 @@ export enum IdentityCheckStep {
   IDENTIFICATION = 'identity-check',
   CONFIRMATION = 'honor-statement',
   END = 'end',
-}
-
-export type SubscriptionScreen = keyof SubscriptionRootStackParamList
-
-export type StepConfig<T> = {
-  name: T
-  icon: Record<StepButtonState, React.FC<IconInterface>>
-  firstScreen: SubscriptionScreen
-}
-
-export type StepDetails<T> = StepConfig<T> & {
-  title: string
-  subtitle?: string
-  stepState: StepButtonState
-}
-
-export enum StepButtonState {
-  'COMPLETED' = 'completed',
-  'CURRENT' = 'current',
-  'DISABLED' = 'disabled',
-  'RETRY' = 'retry',
 }
 
 export interface RehydrationProfile {
